+++ demos/src/Extensions/DragHandleWithNodeViews/React/extensions/recommendation/Recommendation.jsx
@@ -58,7 +58,9 @@ export const Recommendation = Node.create({
   },
 
   addNodeView() {
-    return ReactNodeViewRenderer(RecommendationView)
+    return ReactNodeViewRenderer(RecommendationView, {
+      trackNodeViewPosition: true,
+    })
   },
 })