+++ packages/extension-drag-handle-react/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @tiptap/extension-drag-handle-react
 
+## 3.23.5
+
+### Patch Changes
+
+- ec1afc3: Fix missing forwarding of getReferencedVirtualElement in DragHandle React component
+- Updated dependencies [b5f34fc]
+- Updated dependencies [95e138c]
+  - @tiptap/react@3.23.5
+  - @tiptap/extension-drag-handle@3.23.5
+  - @tiptap/pm@3.23.5
+
 ## 3.23.4
 
 ### Patch Changes