+++ packages/extension-drag-handle-vue-3/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @tiptap/extension-drag-handle-vue-3
 
+## 3.23.5
+
+### Patch Changes
+
+- Updated dependencies [95e138c]
+  - @tiptap/vue-3@3.23.5
+  - @tiptap/extension-drag-handle@3.23.5
+  - @tiptap/pm@3.23.5
+
 ## 3.23.4
 
 ### Patch Changes