+++ packages/extension-file-handler/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @tiptap/extension-file-handler
 
+## 3.23.5
+
+### Patch Changes
+
+- Updated dependencies [835caf5]
+- Updated dependencies [95e138c]
+  - @tiptap/core@3.23.5
+  - @tiptap/extension-text-style@3.23.5
+  - @tiptap/pm@3.23.5
+
 ## 3.23.4
 
 ### Patch Changes