+++ packages/extension-node-range/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @tiptap/extension-node-range
 
+## 3.23.5
+
+### Patch Changes
+
+- Updated dependencies [835caf5]
+- Updated dependencies [95e138c]
+  - @tiptap/core@3.23.5
+  - @tiptap/pm@3.23.5
+
 ## 3.23.4
 
 ### Patch Changes